+++ AdyenActions/AdyenActionComponent.swift
@@ -202,34 +202,19 @@ public final class AdyenActionComponent: ActionComponent, ActionHandlingComponen
         component.handle(action)
         currentActionComponent = component
     }
-    
+
     private func handle(_ action: QRCodeAction) {
-        
-        switch action.paymentMethodType {
-        case .pix:
-            let component = PixActionComponent(context: context)
-            component.configuration.style = configuration.style.qrCodeComponentStyle
-            component._isDropIn = _isDropIn
-            component.delegate = delegate
-            component.presentationDelegate = presentationDelegate
-            component.configuration.localizationParameters = configuration.localizationParameters
+        let component = QRCodeActionComponent(context: context)
+        component.configuration.style = configuration.style.qrCodeComponentStyle
+        component._isDropIn = _isDropIn
+        component.delegate = delegate
+        component.presentationDelegate = presentationDelegate
+        component.configuration.localizationParameters = configuration.localizationParameters
 
-            component.handle(action)
-            currentActionComponent = component
-
-        case .promptPay:
-            let component = QRCodeActionComponent(context: context)
-            component.configuration.style = configuration.style.qrCodeComponentStyle
-            component._isDropIn = _isDropIn
-            component.delegate = delegate
-            component.presentationDelegate = presentationDelegate
-            component.configuration.localizationParameters = configuration.localizationParameters
-
-            component.handle(action)
-            currentActionComponent = component
-        }
+        component.handle(action)
+        currentActionComponent = component
     }
-    
+
     private func handle(_ action: DocumentAction) {
         let component = DocumentComponent(context: context)
         component.configuration.style = configuration.style.documentActionComponentStyle
