+++ Tests/Components Tests/QRCode/QRCodeActionComponentTests.swift
@@ -8,7 +8,7 @@
 @testable import AdyenActions
 import XCTest
 
-class PixActionComponentTests: XCTestCase {
+class QRCodeActionComponentTests: XCTestCase {
 
     var context: AdyenContext!
 
@@ -23,86 +23,8 @@ class PixActionComponentTests: XCTestCase {
     }
 
     lazy var method = InstantPaymentMethod(type: .other("pix"), name: "pix")
-    let payment = Payment(amount: Amount(value: 2, currencyCode: "BRL"), countryCode: "BR")
     let action = QRCodeAction(paymentMethodType: .pix, qrCodeData: "DummyData", paymentData: "DummyData")
     let componentData = ActionComponentData(details: AwaitActionDetails(payload: "DummyPayload"), paymentData: "DummyData")
-    
-    func testUIConfiguration() {
-        let dummyExpectation = expectation(description: "Dummy Expectation")
-        var style = QRCodeComponentStyle()
-        
-        style.copyButton = ButtonStyle(
-            title: TextStyle(font: .preferredFont(forTextStyle: .callout), color: .blue, textAlignment: .justified),
-            cornerRadius: 4,
-            background: .black
-        )
-        
-        style.instructionLabel = TextStyle(
-            font: .systemFont(ofSize: 20, weight: .semibold),
-            color: .darkGray,
-            textAlignment: .left
-        )
-        
-        style.progressView = ProgressViewStyle(
-            progressTintColor: .cyan, trackTintColor: .brown
-        )
-        
-        style.expirationLabel = TextStyle(
-            font: .boldSystemFont(ofSize: 25),
-            color: .blue, textAlignment: .right
-        )
-        
-        style.logoCornerRounding = .fixed(10)
-        
-        style.backgroundColor = UIColor.Adyen.componentSeparator
-        
-        let sut = PixActionComponent(context: context)
-        sut.configuration.style = style
-        let presentationDelegate = PresentationDelegateMock()
-        sut.presentationDelegate = presentationDelegate
-        
-        presentationDelegate.doPresent = { [weak self] component in
-            XCTAssertNotNil(component.viewController as? PixViewController)
-            let viewController = component.viewController as! PixViewController
-            
-            UIApplication.shared.keyWindow?.rootViewController = viewController
-            
-            self?.wait(for: .milliseconds(300))
-            
-            let copyButton: SubmitButton? = viewController.view.findView(by: "copyButton")
-            let instructionLabel: UILabel? = viewController.view.findView(by: "instructionLabel")
-            let progressView: UIProgressView? = viewController.view.findView(by: "progressView")
-            let expirationLabel: UILabel? = viewController.view.findView(by: "expirationLabel")
-            let logo: UIImageView? = viewController.view.findView(by: "logo")
-            
-            // Test copy button
-            XCTAssertEqual(copyButton?.backgroundColor, style.copyButton.backgroundColor)
-            XCTAssertEqual(copyButton?.layer.cornerRadius, 4)
-            
-            // Test instruction label
-            XCTAssertEqual(instructionLabel?.font, style.instructionLabel.font)
-            XCTAssertEqual(instructionLabel?.textColor, style.instructionLabel.color)
-            XCTAssertEqual(instructionLabel?.textAlignment, style.instructionLabel.textAlignment)
-            
-            // Test progress view
-            XCTAssertEqual(progressView?.progressTintColor, style.progressView.progressTintColor)
-            XCTAssertEqual(progressView?.trackTintColor, style.progressView.trackTintColor)
-            
-            // Test expiration label
-            XCTAssertEqual(expirationLabel?.font, style.expirationLabel.font)
-            XCTAssertEqual(expirationLabel?.textColor, style.expirationLabel.color)
-            XCTAssertEqual(expirationLabel?.textAlignment, style.expirationLabel.textAlignment)
-            
-            // Test logo
-            XCTAssertEqual(logo?.layer.cornerRadius, 10)
-            
-            dummyExpectation.fulfill()
-        }
-        
-        sut.handle(action)
-        
-        waitForExpectations(timeout: 10, handler: nil)
-    }
     
     func testComponentTimeout() {
         let dummyExpectation = expectation(description: "Dummy Expectation")
@@ -121,7 +43,7 @@ class PixActionComponentTests: XCTestCase {
             }
         )
         
-        let sut = PixActionComponent(context: context,
+        let sut = QRCodeActionComponent(context: context,
                                   pollingComponentBuilder: builder,
                                   timeoutInterval: 2.0)
         let componentDelegate = ActionComponentDelegateMock()
@@ -137,8 +59,8 @@ class PixActionComponentTests: XCTestCase {
         
         let presentationDelegate = PresentationDelegateMock()
         presentationDelegate.doPresent = { component in
-            XCTAssertNotNil(component.viewController as? PixViewController)
-            let viewController = component.viewController as! PixViewController
+            XCTAssertNotNil(component.viewController as? QRCodeViewController)
+            let viewController = component.viewController as! QRCodeViewController
             
             UIApplication.shared.keyWindow?.rootViewController = viewController
         }
@@ -163,7 +85,7 @@ class PixActionComponentTests: XCTestCase {
             }
         )
         
-        let sut = PixActionComponent(context: context,
+        let sut = QRCodeActionComponent(context: context,
                                   pollingComponentBuilder: builder,
                                   timeoutInterval: 2.0)
         
@@ -184,8 +106,8 @@ class PixActionComponentTests: XCTestCase {
         
         let presentationDelegate = PresentationDelegateMock()
         presentationDelegate.doPresent = { component in
-            XCTAssertNotNil(component.viewController as? PixViewController)
-            let viewController = component.viewController as! PixViewController
+            XCTAssertNotNil(component.viewController as? QRCodeViewController)
+            let viewController = component.viewController as! QRCodeViewController
             
             UIApplication.shared.keyWindow?.rootViewController = viewController
         }
@@ -210,7 +132,7 @@ class PixActionComponentTests: XCTestCase {
             }
         )
         
-        let sut = PixActionComponent(context: context,
+        let sut = QRCodeActionComponent(context: context,
                                   pollingComponentBuilder: builder,
                                   timeoutInterval: 2.0)
         
@@ -231,8 +153,8 @@ class PixActionComponentTests: XCTestCase {
         
         let presentationDelegate = PresentationDelegateMock()
         presentationDelegate.doPresent = { component in
-            XCTAssertNotNil(component.viewController as? PixViewController)
-            let viewController = component.viewController as! PixViewController
+            XCTAssertNotNil(component.viewController as? QRCodeViewController)
+            let viewController = component.viewController as! QRCodeViewController
             
             UIApplication.shared.keyWindow?.rootViewController = viewController
         }
@@ -245,31 +167,60 @@ class PixActionComponentTests: XCTestCase {
         waitForExpectations(timeout: 10, handler: nil)
     }
     
-    func testCopyButton() {
+    func testCopyCodeButton() {
         let dummyExpectation = expectation(description: "Dummy Expectation")
-        
-        let sut = PixActionComponent(context: context)
+
+        let sut = QRCodeActionComponent(context: context)
         let presentationDelegate = PresentationDelegateMock()
         sut.presentationDelegate = presentationDelegate
-        
+
         presentationDelegate.doPresent = { [self] component in
-            XCTAssertNotNil(component.viewController as? PixViewController)
-            let viewController = component.viewController as! PixViewController
-            
+            XCTAssertNotNil(component.viewController as? QRCodeViewController)
+            let viewController = component.viewController as! QRCodeViewController
+
             UIApplication.shared.keyWindow?.rootViewController = viewController
-            
+
             wait(for: .milliseconds(300))
-            
-            let copyButton: SubmitButton? = viewController.view.findView(by: "copyButton")
+
+            let copyButton: SubmitButton? = viewController.view.findView(by: "copyCodeButton")
             XCTAssertNotNil(copyButton)
             copyButton?.sendActions(for: .touchUpInside)
-            
+
             XCTAssertEqual(self.action.qrCodeData, UIPasteboard.general.string)
-            
+
             dummyExpectation.fulfill()
         }
-        
+
         sut.handle(action)
         waitForExpectations(timeout: 10, handler: nil)
     }
+    
+    func testSaveAsImageButton() {
+        lazy var method = InstantPaymentMethod(type: .other("promptpay"), name: "promptpay")
+        let action = QRCodeAction(paymentMethodType: .promptPay, qrCodeData: "DummyData", paymentData: "DummyData")
+
+        let dummyExpectation = expectation(description: "Dummy Expectation")
+
+        let sut = QRCodeActionComponent(context: context)
+        let presentationDelegate = PresentationDelegateMock()
+        sut.presentationDelegate = presentationDelegate
+        let delgate = QRCodeViewDelegateMock()
+
+        presentationDelegate.doPresent = { [self] component in
+            XCTAssertNotNil(component.viewController as? QRCodeViewController)
+            let viewController = component.viewController as! QRCodeViewController
+            UIApplication.shared.keyWindow?.rootViewController = viewController
+            viewController.qrCodeView.delegate = delgate
+            wait(for: .milliseconds(300))
+            let saveAsImageButton: SubmitButton? = viewController.view.findView(by: "saveAsImageButton")
+            XCTAssertNotNil(saveAsImageButton)
+            saveAsImageButton?.sendActions(for: .touchUpInside)
+            XCTAssertTrue(delgate.saveAsImageCalled)
+            dummyExpectation.fulfill()
+        }
+
+        sut.handle(action)
+        waitForExpectations(timeout: 10, handler: nil)
+    }
+
 }
