+++ Tests/Components Tests/QRCode/QRCodeViewDelegateMock.swift
@@ -10,13 +10,22 @@
 final class QRCodeViewDelegateMock: QRCodeViewDelegate {
 
     var saveAsImageCallsCount = 0
+    var copyToPasteboardCount = 0
 
     var saveAsImageCalled: Bool {
         saveAsImageCallsCount > 0
     }
 
+    var copyToPasteboardCalled: Bool {
+        copyToPasteboardCount > 0
+    }
+
     func saveAsImage(qrCodeImage: UIImage?, sourceView: UIView) {
         saveAsImageCallsCount += 1
+    }
+
+    func copyToPasteboard(with action: QRCodeAction) {
+        copyToPasteboardCount += 1
     }
 
 }
