+++ Tests/Session Tests/SessionTests.swift
@@ -639,7 +639,7 @@ class SessionTests: XCTestCase {
         
         let paymentMethod = expectedPaymentMethods.regular.first as! GiftCardPaymentMethod
         let paymentComponent = PaymentComponentMock(paymentMethod: paymentMethod)
-        let actionComponent = PixActionComponent(context: context)
+        let actionComponent = QRCodeActionComponent(context: context)
         
         let didFailExpectation = expectation(description: "didFail should be called")
         sessionDelegate.onDidFail = { (error, component, session) in
@@ -691,7 +691,7 @@ class SessionTests: XCTestCase {
         )
         
         dropIn.didFail(with: ComponentError.paymentMethodNotSupported, from: paymentComponent)
-        dropIn.didOpenExternalApplication(component: PixActionComponent(context: context))
+        dropIn.didOpenExternalApplication(component: QRCodeActionComponent(context: context))
         dropIn.didSubmit(paymentData, from: paymentComponent)
         dropIn.didProvide(actionData, from: actionComponent)
         sut.sessionContext.resultCode = .authorised
@@ -727,7 +727,7 @@ class SessionTests: XCTestCase {
             ),
             paymentData: "payment_data"
         )
-        sut.didProvide(actionData, from: PixActionComponent(context: context))
+        sut.didProvide(actionData, from: QRCodeActionComponent(context: context))
         wait(for: [didCompleteExpectation], timeout: 2)
     }
     
@@ -758,7 +758,7 @@ class SessionTests: XCTestCase {
             ),
             paymentData: "payment_data"
         )
-        sut.didProvide(actionData, from: PixActionComponent(context: context))
+        sut.didProvide(actionData, from: QRCodeActionComponent(context: context))
         wait(for: [didCompleteExpectation], timeout: 2)
     }
     
@@ -789,7 +789,7 @@ class SessionTests: XCTestCase {
             ),
             paymentData: "payment_data"
         )
-        sut.didProvide(actionData, from: PixActionComponent(context: context))
+        sut.didProvide(actionData, from: QRCodeActionComponent(context: context))
         wait(for: [didCompleteExpectation], timeout: 2)
     }
     
@@ -820,7 +820,7 @@ class SessionTests: XCTestCase {
             ),
             paymentData: "payment_data"
         )
-        sut.didProvide(actionData, from: PixActionComponent(context: context))
+        sut.didProvide(actionData, from: QRCodeActionComponent(context: context))
         wait(for: [didCompleteExpectation], timeout: 2)
     }
     
@@ -851,7 +851,7 @@ class SessionTests: XCTestCase {
             ),
             paymentData: "payment_data"
         )
-        sut.didProvide(actionData, from: PixActionComponent(context: context))
+        sut.didProvide(actionData, from: QRCodeActionComponent(context: context))
         wait(for: [didCompleteExpectation], timeout: 2)
     }
     
@@ -917,7 +917,7 @@ class SessionTests: XCTestCase {
             ),
             paymentData: "payment_data"
         )
-        sut.didProvide(actionData, from: PixActionComponent(context: context))
+        sut.didProvide(actionData, from: QRCodeActionComponent(context: context))
         wait(for: [didCompleteExpectation], timeout: 2)
     }
     
@@ -948,7 +948,7 @@ class SessionTests: XCTestCase {
             ),
             paymentData: "payment_data"
         )
-        sut.didProvide(actionData, from: PixActionComponent(context: context))
+        sut.didProvide(actionData, from: QRCodeActionComponent(context: context))
         wait(for: [didCompleteExpectation], timeout: 2)
     }
